src/elastic-search/requirements.txt
elasticsearch==8.13.0
python-dotenv==1.1.0
